combine with cli.ts, inbox.ts, main.ts
